requirements.txt
robotframework==6.1.1
robotframework-browser==18.0.0
robotframework-pythonlibcore>=4.3.0,<5.0.0
python-dotenv==1.0.0

# Windows-specific dependencies (automatically skipped on other platforms)
nvda-automation==0.2.0;platform_system=="Windows"
pywinauto==0.6.8;platform_system=="Windows" 